Minor fixes.

+++ Classes/SPSuperSFV.h
@@ -76,9 +76,9 @@ typedef unsigned char u8;
 - (void)removeSelectedRecords:(id) sender;
 - (void)fileAddingThread;
 - (void)addFiles:(NSTimer *)timer;
-- (void)didEndSaveSheet:(NSSavePanel *)savePanel returnCode:(int)returnCode contextInfo:(void *)contextInfo;
-- (void)didEndOpenSheet:(NSOpenPanel *)openPanel returnCode:(int)returnCode contextInfo:(void *)contextInfo;
-- (void)didEndRemoveAllSheet:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)didEndSaveSheet:(NSSavePanel *)savePanel returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
+- (void)didEndOpenSheet:(NSOpenPanel *)openPanel returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
+- (void)didEndRemoveAllSheet:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 @property (readonly, copy) NSString *_applicationVersion;
 
 - (void)sortWithDescriptor:(NSSortDescriptor*)descriptor;
